addressed changes requested from PR review
