+++ cypress/e2e/po/edit/provisioning.cattle.io.cluster/create/cluster-create-gke.po.ts
@@ -1,6 +1,7 @@
 import PagePo from '@/cypress/e2e/po/pages/page.po';
 import ClusterManagerCreatePagePo from '@/cypress/e2e/po/edit/provisioning.cattle.io.cluster/create/cluster-create.po';
 import LabeledInputPo from '@/cypress/e2e/po/components/labeled-input.po';
+import LabeledSelectPo from '@/cypress/e2e/po/components/labeled-select.po';
 import GKECloudCredentialsCreateEditPo from '@/cypress/e2e/po/edit/cloud-credentials-gke.po';
 import AsyncButtonPo from '@/cypress/e2e/po/components/async-button.po';
 
@@ -30,5 +31,13 @@ export default class ClusterManagerCreateGKEPagePo extends ClusterManagerCreateP
 
   saveCreateGkeCluster(): AsyncButtonPo {
     return new AsyncButtonPo('[data-testid="form-save"]', this.self());
+  }
+
+  static getGkeVersionSelect() {
+    return new LabeledSelectPo('[data-testid="gke-version-select"]');
+  }
+
+  static getGkeZoneSelect() {
+    return new LabeledSelectPo('[data-testid="gke-zone-select"]');
   }
 }
